verilog: Don't increment twice counters.
The wb_active (so both strobe and active) can cover _two_ clock
cycles. Which meant that we would get to this function twice
and increment the internal index twice.

Not good. Lets just depend on 'transmit' value which is only
set once (after we have processed the write command) to guard
against that.

Also add some extra logic in case we do screw up somehow
the indexing and we set the panic reg.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>