verilog/test: Fix CTRL_SHA1_DIGEST to read correctly.
Prior to this patch we would increase the idx every
clock cycle as long as wb_active is on. That interestingly
covers _two_ clock cycles so we ended up increasing
the idx twice in one wb_active pulse.

That is wrong, we should be increasing it once per read, not
per clock cycle. As such lets depend on the behavior
that transmit is not yet set (as we set it _after_ we have
read the first time), and do our incrementation then.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>