sha1: Fix state transition.
We would do the state transition a bit too late (or early).

That is we need the updated state and k value at the start
of the compute value being high. We increment the index
value every two cycles:

 - The first cycle the copy_values is high (so copying
   older values), and the 'temp' has the value from the
   prior index. So:
   copy_values=1
   compute=0
   index=19

 - The next cycle the compute is high and we do the operation.
   This cycle is also when we increase the index counter.
   copy_values=0
   compute=1
   index=19 <= this changes on the next cycle count

And then we go in the "next" index value. The index had been
registered in the prior cycle, so:
   index=20
   temp=<value computed using prior w[index]>
   copy_values=1
   compute=0
   (So we copy the values over).
 - Next cycle the compute is high .. and so on.

The problem we had was that the 'state' changed when index (at 19)
registered (and copy_values was high with compute low) which meant
we would change the state (and k value) using the next LOOP function
on the next cycle (when compute was high).

Depending on compute being heigh solves the problem.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>