test: Expand the test-case to print values at the right
time. That is the 'temp' value should be printed when
the index has moved forward; The w[i] should be printed
only within 0->79 index and no more. And the state
(now) registers at the index having changed.

While at it - also make that whole thing be capable
of running a couple of times.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>