test: Wrap in the TB the wrapper_sha1
.. instead of having it standalone. That allows us
to be able to reference the 'status' register.

But it also uncovered some other issues in the code such
as the test-cases not being prepared to be run from
a module higher than wrapper_sha1 and needing the
logic to handle that.

The GL detection logic is not really working right either.
It will always set GL to true when running with the dump_wrapper.v
because the power1,power2 are in it (even if using the LVS
as well both of those are used). Will have to be reworked.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>